# Expose the `InitialLevel.IgnoreChanges()` optimization