--- a/src/SerilogTracing/ActivityListenerConfiguration.cs
+++ b/src/SerilogTracing/ActivityListenerConfiguration.cs
@@ -93,7 +93,7 @@ public class ActivityListenerConfiguration
     /// <returns>A handle that must be kept alive while tracing is required, and disposed afterwards.</returns>
     public IDisposable TraceTo(ILogger logger)
     {
-        return LoggerActivityListener.Configure(this, () => logger, ignoreLevelChanges: false);
+        return LoggerActivityListener.Configure(this, () => logger);
     }
 
     /// <summary>
@@ -105,6 +105,6 @@ public class ActivityListenerConfiguration
     /// <returns>A handle that must be kept alive while tracing is required, and disposed afterwards.</returns>
     public IDisposable TraceToSharedLogger()
     {
-        return LoggerActivityListener.Configure(this, () => Log.Logger, ignoreLevelChanges: false);
+        return LoggerActivityListener.Configure(this, () => Log.Logger);
     }
 }