#84 introduced an internal setting `ignoreLevelChanges` that optimizes the way the SerilogTracing activity listener handles external sources.

When the setting is enabled, external sources that are suppressed are permanently ignored, an no further checks are made when they attempt to start activities.

This is in contrast to the default behavior, which performs a check every time an activity is started, to properly respond to minimum level changes in the target logger.

This PR exposes the setting as `InitialLevel.IgnoreChanges()`:

```csharp
        _loggerActivityListener = new ActivityListenerConfiguration()
            .InitialLevel.IgnoreChanges()
            .InitialLevel.Override(_ignoredSource.Name, LogEventLevel.Verbose)
            .Instrument.WithDefaultInstrumentation(false)
            .TraceTo(_log);
```

Benchmark results are tricky to interpret, because this benchmark isn't particularly well-constructed, but they're a reasonable indication of the expected performance gains:

```
BenchmarkDotNet v0.13.12, macOS Sonoma 14.6.1 (23G93) [Darwin 23.6.0]
Apple M3 Pro, 1 CPU, 11 logical and 11 physical cores
.NET SDK 8.0.402
  [Host]     : .NET 8.0.8 (8.0.824.36612), Arm64 RyuJIT AdvSIMD
  DefaultJob : .NET 8.0.8 (8.0.824.36612), Arm64 RyuJIT AdvSIMD
```

**Before:**

| Method                 | Mean     | Error   | StdDev  | Ratio | RatioSD | Gen0   | Gen1   | Allocated | Alloc Ratio |
|----------------------- |---------:|--------:|--------:|------:|--------:|-------:|-------:|----------:|------------:|
| ActivitySourceBaseline | 202.9 ns | 0.90 ns | 0.80 ns |  1.00 |    0.00 | 0.0823 |      - |     688 B |        1.00 |
| ActivityListenerOnly   | 347.6 ns | 1.63 ns | 1.45 ns |  1.71 |    0.01 | 0.1469 |      - |    1232 B |        1.79 |
| StartThenDispose       | 359.9 ns | 1.56 ns | 1.46 ns |  1.77 |    0.01 | 0.1392 | 0.0005 |    1168 B |        1.70 |
| StartThenComplete      | 727.8 ns | 1.24 ns | 1.16 ns |  3.59 |    0.02 | 0.2394 | 0.0010 |    2008 B |        2.92 |

**After:**

| Method                 | Mean     | Error   | StdDev  | Ratio | Gen0   | Allocated | Alloc Ratio |
|----------------------- |---------:|--------:|--------:|------:|-------:|----------:|------------:|
| ActivitySourceBaseline | 141.4 ns | 0.34 ns | 0.32 ns |  1.00 | 0.0496 |     416 B |        1.00 |
| ActivityListenerOnly   | 319.7 ns | 0.77 ns | 0.68 ns |  2.26 | 0.1307 |    1096 B |        2.63 |
| StartThenDispose       | 306.1 ns | 0.54 ns | 0.51 ns |  2.17 | 0.1230 |    1032 B |        2.48 |
| StartThenComplete      | 659.5 ns | 1.12 ns | 0.94 ns |  4.67 | 0.2232 |    1872 B |        4.50 |

The two `SerilogTracing` cases - `StartThenDispose()` and `StartThenComplete()` - pick up a better than 10% improvement in execution time, a similar allocation benefit, and both stop triggering Gen1 collections.

The `Activity*` cases point to how external activity sources will be made more efficient with this change - a better than 25% improvement in run-time and allocations. This is the case that's motivating my PR - I'm integrating SerilogTracing into a codebase with extremely high-frequency but suppressed activities present.